requirements.txt
# Core
streamlit>=1.32.0
fastapi>=0.110.0
uvicorn>=0.27.0

# Database
psycopg2-binary>=2.9.9

# RAG Pipeline
chromadb>=0.4.22
sentence-transformers>=2.5.1
ollama>=0.1.7

# PDF Generation
reportlab>=4.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
requests>=2.31.0
PyYAML>=6.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.2.0
httpx>=0.27.0
